lib/db: Move database schema migration into its own file (#4985)
